chore(utils): group reports related entities in a separate folder
