+++ packages/utils/perf/score-report/index.ts
@@ -1,6 +1,6 @@
 import * as Benchmark from 'benchmark';
 import { Report } from '@code-pushup/models';
-import { scoreReport } from '../../src/lib/scoring';
+import { scoreReport } from '../../src/lib/reports/scoring';
 import { scoreReportOptimized0 } from './optimized0';
 import { scoreReportOptimized1 } from './optimized1';
 import { scoreReportOptimized2 } from './optimized2';
