+++ packages/utils/src/index.ts
@@ -37,8 +37,11 @@ export {
   isPromiseRejectedResult,
 } from './lib/guards';
 export { logMultipleResults } from './lib/log-results';
-export { NEW_LINE } from './lib/md';
 export { ProgressBar, getProgressBar } from './lib/progress';
+export { reportToMd } from './lib/reports/report-to-md';
+export { reportToStdout } from './lib/reports/report-to-stdout';
+export { ScoredReport, scoreReport } from './lib/reports/scoring';
+export { sortReport } from './lib/reports/sorting';
 export {
   CODE_PUSHUP_DOMAIN,
   FOOTER_PREFIX,
@@ -46,10 +49,7 @@ export {
   calcDuration,
   compareIssueSeverity,
   loadReport,
-} from './lib/report';
-export { reportToMd } from './lib/report-to-md';
-export { reportToStdout } from './lib/report-to-stdout';
-export { ScoredReport, scoreReport } from './lib/scoring';
+} from './lib/reports/utils';
 export {
   CliArgsObject,
   countOccurrences,
@@ -62,4 +62,3 @@ export {
   toUnixPath,
 } from './lib/transform';
 export { verboseUtils } from './lib/verbose-utils';
-export { sortReport } from './lib/sort-report';
