--- packages/utils/src/lib/md/constants.ts
@@ -1,1 +1,0 @@
-export const NEW_LINE = '\n';
