+++ packages/utils/src/lib/reports/constants.ts
@@ -1,3 +1,5 @@
+export const NEW_LINE = '\n';
+
 export const SCORE_COLOR_RANGE = {
   GREEN_MIN: 0.9,
   YELLOW_MIN: 0.5,
