+++ packages/utils/src/lib/reports/md/details.ts
