+++ packages/utils/src/lib/reports/md/index.ts
@@ -1,7 +1,6 @@
 export * from './details';
-export * from './headline';
-export * from './table';
-export * from './constants';
 export * from './font-style';
+export * from './headline';
 export * from './link';
 export * from './list';
+export * from './table';
