+++ packages/utils/src/lib/reports/report-to-stdout.integration.test.ts
@@ -2,7 +2,7 @@ import { describe } from 'vitest';
 import { reportMock } from '@code-pushup/testing-utils';
 import { reportToStdout } from './report-to-stdout';
 import { scoreReport } from './scoring';
-import { sortReport } from './sort-report';
+import { sortReport } from './sorting';
 
 describe('report-to-stdout', () => {
   it('should contain all sections when using the fixture report', () => {
